requirements.txt
openai>=1.40.0
fastapi>=0.112.0
uvicorn>=0.30.0
pydantic>=2.7.0
PyYAML>=6.0.1
requests>=2.32.0
python-dotenv>=1.0.0